Se actualiza la validación de las imagenes si viene vacio o null

+++ app/Http/Controllers/Controlador.php
@@ -129,7 +129,8 @@ class Controlador extends Controller
                     if($request->hasFile('image')){
                         $probar = $request->file('image')->getClientOriginalName();
                         $imagenes = $request->file('image')->storeAs("public/archivo_imagenes/". $id->id,$probar);
-                        if($id->image != ''){
+                        print_r(public_path($id->image));
+                        if($id->image != null && $id->image != ''){
                             unlink(public_path($id->image));
                         }
                         $image = Storage::url($imagenes);
